pyproject.toml
[project]
name = "greenhouse-mcp-server"
version = "0.1.0"
description = "MCP server for Greenhouse ATS talent acquisition"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "httpx",
    "python-dotenv",
    "PyJWT",
    "pymupdf",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
